Avoid PHP warning if 'muplugins_loaded' key is missing
I've observed one of our HM Altis sites filling up its logs with warnings around 
"Undefined index: muplugins_loaded" when evaluating this check. The isset()
check on its own is sufficient to check whether we are in the state this
conditional is meant to address, and removing the is_array sidesteps triggering
a warning if the filters are not yet set up.