Don't always set the `machine_name`
I wanted to set the `machine_name` in my `composer.json` file however I couldn't as it's set to `$config['hosts'][0]`. I understand the logic for this would be so that you don't have duplicate machines named `default` from the upstream Chassis default. However, I'd like to have the option to set it so this should take care of that.